+++ src/main/java/ar/edu/unrc/game2048/experiments/configurations/INeuralNetworkInterfaceFor2048.java
@@ -23,9 +23,7 @@ import ar.edu.unrc.coeus.tdlearning.interfaces.IAction;
 import ar.edu.unrc.coeus.tdlearning.learning.ELearningStyle;
 import ar.edu.unrc.coeus.tdlearning.learning.TDLambdaLearning;
 import ar.edu.unrc.game2048.Action;
-import ar.edu.unrc.game2048.EncogConfiguration2048;
 import ar.edu.unrc.game2048.Game2048;
-import ar.edu.unrc.game2048.NTupleConfiguration2048;
 
 import java.io.File;
 import java.io.OutputStream;
