+++ src/main/java/ar/edu/unrc/game2048/experiments/configurations/librariesinterfaces/EncogExperimentInterface.java
@@ -20,7 +20,7 @@ package ar.edu.unrc.game2048.experiments.configurations.librariesinterfaces;
 
 import ar.edu.unrc.coeus.interfaces.INeuralNetworkInterface;
 import ar.edu.unrc.coeus.utils.FunctionUtils;
-import ar.edu.unrc.game2048.EncogConfiguration2048;
+import ar.edu.unrc.game2048.experiments.configurations.EncogConfiguration2048;
 import ar.edu.unrc.game2048.experiments.configurations.INeuralNetworkInterfaceFor2048;
 import org.encog.engine.network.activation.ActivationFunction;
 import org.encog.engine.network.activation.ActivationLinear;
