+++ src/main/java/ar/edu/unrc/game2048/experiments/configurations/ntuples/ConfigNTupleBasicTanH_32768.java
@@ -21,8 +21,8 @@ package ar.edu.unrc.game2048.experiments.configurations.ntuples;
 import ar.edu.unrc.coeus.tdlearning.training.ntuple.SamplePointValue;
 import ar.edu.unrc.coeus.utils.FunctionUtils;
 import ar.edu.unrc.game2048.GameBoard;
-import ar.edu.unrc.game2048.NTupleConfiguration2048;
 import ar.edu.unrc.game2048.Tile;
+import ar.edu.unrc.game2048.experiments.configurations.NTupleConfiguration2048;
 import org.encog.util.arrayutil.NormalizationAction;
 import org.encog.util.arrayutil.NormalizedField;
