+++ src/main/java/ar/edu/unrc/game2048/experiments/performance/ConcurrencyExperiment_Basic.java
@@ -21,7 +21,7 @@ package ar.edu.unrc.game2048.experiments.performance;
 import ar.edu.unrc.coeus.interfaces.INeuralNetworkInterface;
 import ar.edu.unrc.coeus.tdlearning.learning.TDLambdaLearning;
 import ar.edu.unrc.coeus.tdlearning.training.ntuple.NTupleSystem;
-import ar.edu.unrc.game2048.EncogConfiguration2048;
+import ar.edu.unrc.game2048.experiments.configurations.EncogConfiguration2048;
 import ar.edu.unrc.game2048.experiments.configurations.LearningExperiment;
 import ar.edu.unrc.game2048.experiments.configurations.librariesinterfaces.EncogExperimentInterface;
 import ar.edu.unrc.game2048.experiments.configurations.perceptrons.ConfigPerceptronNTupleTanHSimplified_512;
