+++ src/test/java/ar/edu/unrc/game2048/experiments/configurations/librariesinterfaces/EncogExperimentInterfaceTest.java
@@ -20,9 +20,9 @@ package ar.edu.unrc.game2048.experiments.configurations.librariesinterfaces;
 
 import ar.edu.unrc.coeus.interfaces.INeuralNetworkInterface;
 import ar.edu.unrc.coeus.utils.FunctionUtils;
-import ar.edu.unrc.game2048.EncogConfiguration2048;
 import ar.edu.unrc.game2048.Game2048;
 import ar.edu.unrc.game2048.GameBoard;
+import ar.edu.unrc.game2048.experiments.configurations.EncogConfiguration2048;
 import org.encog.engine.network.activation.ActivationFunction;
 import org.encog.engine.network.activation.ActivationLinear;
 import org.encog.engine.network.activation.ActivationSigmoid;
